[Packetbeat] Create x-pack magefile (#21979)
* Create packetbeat x-pack magefile

* add changelog entry

* Fix up packetbeat magefile

* Add makefile for x-pack directory

* Uncomment packaging

* Add desired platforms to Jenkinsfile

* Add back in fields yml and kibana dashboards on package

* Add generated configs in x-pack/packetbeat

* Suggested changes

(cherry picked from commit 3e73c99391799badcb5b5e8e63eba779928d2426)
