+++ x-pack/packetbeat/Makefile
@@ -1,0 +1,3 @@
+ES_BEATS ?= ../..
+
+include $(ES_BEATS)/dev-tools/make/mage.mk
